updated plot function
